requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
djangorestframework==3.13.1
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.16.0
packaging==21.3
Pillow==9.1.1
psycopg2-binary==2.9.3
Pygments==2.12.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0

celery~=5.2.7